latona: fixup UART_SEL gpio
OMAP_GPIO_UART_SEL in Latona is used to dynamically switch UART
between the Application Processor (AP) and the Cellular Processor (CP).
The GPIO is pulled high if the AP reserves the UART or is pulled low
when the UART is reserved by the CP. We currently don't require the
Baseband logging and hence we pull up the GPIO and prevent the
modemctl from switching it.