mach: latona-power: use GPIO again to monitor full charge
It seems that the inaccurate SOC and VC values are causing
false reporting of Full battery state, which prevents proper
full charging of the device. CHG_EN GPIO was being used in
the 2.6 Samsung Kernel too. Assuming that the max8855 charger
would prevent overcharging, we could safely bring back the
GPIO check.

Also adjusted the full charge VC to match the 2.6 kernel.